+++ src/engine/collision/collision_system.h
@@ -9,7 +9,7 @@ class CollisionSystem : public LogicSystem
 public:
   CollisionSystem(Context* ecsContext);
 
-  virtual void Update(const double dt) override;
+  virtual void Update(const float dt) override;
 
 private:
   Group* m_CircleComponentGroup;
