+++ src/game/systems/asteroids_spawn_system.h
@@ -12,7 +12,7 @@ class AsteroidsSpawnSystem : public LogicSystem
 public:
   AsteroidsSpawnSystem(Context* ecsContext);
 
-  virtual void Update(const double dt) override;
+  virtual void Update(const float dt) override;
 
 private:
   Group* m_IdentityGroup;
