+++ src/game/systems/player_respawn_system.h
@@ -12,7 +12,7 @@ class PlayerRespawnSystem : public LogicSystem
 public:
   PlayerRespawnSystem(Context* ecsContext);
 
-  virtual void Update(const double dt) override;
+  virtual void Update(const float dt) override;
 
 private:
   Group* m_PlayerGroup;
